loaders: commonize file open logic
consolidated file-opening into a general-purpose function
for all loaders. Removed duplicate code to improve
maintainability and readability.

issue: https://github.com/thorvg/thorvg/issues/3901

+++ src/loaders/external_jpg/tvgJpgLoader.h
@@ -43,7 +43,7 @@ private:
 
     tjhandle jpegDecompressor;
     unsigned char* data = nullptr;
-    unsigned long size = 0;
+    uint32_t size = 0;
     bool freeData = false;
 };
